requirements.txt
django==5.2.1
openai==1.78.1
langchain==0.3.25
faiss-cpu==1.11.0
python-dotenv==1.1.0
djangorestframework==3.16.0
langchain-community==0.3.24
langchain-openai==0.3.16
